LPS-55572 delete old files referencing previous name
