+++ webs/mobile-widgets-web/docroot/WEB-INF/service/com/liferay/screens/service/LiferayScreensUserServiceWrapper.java
@@ -59,6 +59,31 @@ public class LiferayScreensUserServiceWrapper
 			arguments);
 	}
 
+	@Override
+	public boolean sendPasswordByEmailAddress(long companyId,
+		java.lang.String emailAddress)
+		throws com.liferay.portal.kernel.exception.PortalException,
+			com.liferay.portal.kernel.exception.SystemException {
+		return _liferayScreensUserService.sendPasswordByEmailAddress(companyId,
+			emailAddress);
+	}
+
+	@Override
+	public boolean sendPasswordByScreenName(long companyId,
+		java.lang.String screenName)
+		throws com.liferay.portal.kernel.exception.PortalException,
+			com.liferay.portal.kernel.exception.SystemException {
+		return _liferayScreensUserService.sendPasswordByScreenName(companyId,
+			screenName);
+	}
+
+	@Override
+	public boolean sendPasswordByUserId(long userId)
+		throws com.liferay.portal.kernel.exception.PortalException,
+			com.liferay.portal.kernel.exception.SystemException {
+		return _liferayScreensUserService.sendPasswordByUserId(userId);
+	}
+
 	/**
 	 * @deprecated As of 6.1.0, replaced by {@link #getWrappedService}
 	 */
