--- webs/mobile-widgets-web/docroot/WEB-INF/src/com/liferay/mobilewidgets/service/base/MobileWidgetsAssetEntryServiceClpInvoker.java
@@ -1,72 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.mobilewidgets.service.base;
-
-import com.liferay.mobilewidgets.service.MobileWidgetsAssetEntryServiceUtil;
-
-import java.util.Arrays;
-
-/**
- * @author José Manuel Navarro
- * @generated
- */
-public class MobileWidgetsAssetEntryServiceClpInvoker {
-	public MobileWidgetsAssetEntryServiceClpInvoker() {
-		_methodName30 = "getBeanIdentifier";
-
-		_methodParameterTypes30 = new String[] {  };
-
-		_methodName31 = "setBeanIdentifier";
-
-		_methodParameterTypes31 = new String[] { "java.lang.String" };
-
-		_methodName34 = "getAssetEntries";
-
-		_methodParameterTypes34 = new String[] {
-				"com.liferay.portlet.asset.service.persistence.AssetEntryQuery",
-				"java.util.Locale"
-			};
-	}
-
-	public Object invokeMethod(String name, String[] parameterTypes,
-		Object[] arguments) throws Throwable {
-		if (_methodName30.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes30, parameterTypes)) {
-			return MobileWidgetsAssetEntryServiceUtil.getBeanIdentifier();
-		}
-
-		if (_methodName31.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes31, parameterTypes)) {
-			MobileWidgetsAssetEntryServiceUtil.setBeanIdentifier((java.lang.String)arguments[0]);
-
-			return null;
-		}
-
-		if (_methodName34.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes34, parameterTypes)) {
-			return MobileWidgetsAssetEntryServiceUtil.getAssetEntries((com.liferay.portlet.asset.service.persistence.AssetEntryQuery)arguments[0],
-				(java.util.Locale)arguments[1]);
-		}
-
-		throw new UnsupportedOperationException();
-	}
-
-	private String _methodName30;
-	private String[] _methodParameterTypes30;
-	private String _methodName31;
-	private String[] _methodParameterTypes31;
-	private String _methodName34;
-	private String[] _methodParameterTypes34;
-}
