--- webs/mobile-widgets-web/docroot/WEB-INF/src/com/liferay/mobilewidgets/service/base/MobileWidgetsDDLRecordServiceBaseImpl.java
@@ -1,419 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.mobilewidgets.service.base;
-
-import com.liferay.mobilewidgets.service.MobileWidgetsDDLRecordService;
-
-import com.liferay.portal.kernel.bean.BeanReference;
-import com.liferay.portal.kernel.bean.IdentifiableBean;
-import com.liferay.portal.kernel.dao.jdbc.SqlUpdate;
-import com.liferay.portal.kernel.dao.jdbc.SqlUpdateFactoryUtil;
-import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.util.InfrastructureUtil;
-import com.liferay.portal.service.BaseServiceImpl;
-import com.liferay.portal.service.persistence.ClassNamePersistence;
-import com.liferay.portal.service.persistence.UserPersistence;
-
-import com.liferay.portlet.dynamicdatalists.service.persistence.DDLRecordPersistence;
-
-import javax.sql.DataSource;
-
-/**
- * Provides the base implementation for the mobile widgets d d l record remote service.
- *
- * <p>
- * This implementation exists only as a container for the default service methods generated by ServiceBuilder. All custom service methods should be put in {@link com.liferay.mobilewidgets.service.impl.MobileWidgetsDDLRecordServiceImpl}.
- * </p>
- *
- * @author José Manuel Navarro
- * @see com.liferay.mobilewidgets.service.impl.MobileWidgetsDDLRecordServiceImpl
- * @see com.liferay.mobilewidgets.service.MobileWidgetsDDLRecordServiceUtil
- * @generated
- */
-public abstract class MobileWidgetsDDLRecordServiceBaseImpl
-	extends BaseServiceImpl implements MobileWidgetsDDLRecordService,
-		IdentifiableBean {
-	/*
-	 * NOTE FOR DEVELOPERS:
-	 *
-	 * Never modify or reference this class directly. Always use {@link com.liferay.mobilewidgets.service.MobileWidgetsDDLRecordServiceUtil} to access the mobile widgets d d l record remote service.
-	 */
-
-	/**
-	 * Returns the mobile widgets asset entry remote service.
-	 *
-	 * @return the mobile widgets asset entry remote service
-	 */
-	public com.liferay.mobilewidgets.service.MobileWidgetsAssetEntryService getMobileWidgetsAssetEntryService() {
-		return mobileWidgetsAssetEntryService;
-	}
-
-	/**
-	 * Sets the mobile widgets asset entry remote service.
-	 *
-	 * @param mobileWidgetsAssetEntryService the mobile widgets asset entry remote service
-	 */
-	public void setMobileWidgetsAssetEntryService(
-		com.liferay.mobilewidgets.service.MobileWidgetsAssetEntryService mobileWidgetsAssetEntryService) {
-		this.mobileWidgetsAssetEntryService = mobileWidgetsAssetEntryService;
-	}
-
-	/**
-	 * Returns the mobile widgets d d l record remote service.
-	 *
-	 * @return the mobile widgets d d l record remote service
-	 */
-	public com.liferay.mobilewidgets.service.MobileWidgetsDDLRecordService getMobileWidgetsDDLRecordService() {
-		return mobileWidgetsDDLRecordService;
-	}
-
-	/**
-	 * Sets the mobile widgets d d l record remote service.
-	 *
-	 * @param mobileWidgetsDDLRecordService the mobile widgets d d l record remote service
-	 */
-	public void setMobileWidgetsDDLRecordService(
-		com.liferay.mobilewidgets.service.MobileWidgetsDDLRecordService mobileWidgetsDDLRecordService) {
-		this.mobileWidgetsDDLRecordService = mobileWidgetsDDLRecordService;
-	}
-
-	/**
-	 * Returns the mobile widgets user remote service.
-	 *
-	 * @return the mobile widgets user remote service
-	 */
-	public com.liferay.mobilewidgets.service.MobileWidgetsUserService getMobileWidgetsUserService() {
-		return mobileWidgetsUserService;
-	}
-
-	/**
-	 * Sets the mobile widgets user remote service.
-	 *
-	 * @param mobileWidgetsUserService the mobile widgets user remote service
-	 */
-	public void setMobileWidgetsUserService(
-		com.liferay.mobilewidgets.service.MobileWidgetsUserService mobileWidgetsUserService) {
-		this.mobileWidgetsUserService = mobileWidgetsUserService;
-	}
-
-	/**
-	 * Returns the counter local service.
-	 *
-	 * @return the counter local service
-	 */
-	public com.liferay.counter.service.CounterLocalService getCounterLocalService() {
-		return counterLocalService;
-	}
-
-	/**
-	 * Sets the counter local service.
-	 *
-	 * @param counterLocalService the counter local service
-	 */
-	public void setCounterLocalService(
-		com.liferay.counter.service.CounterLocalService counterLocalService) {
-		this.counterLocalService = counterLocalService;
-	}
-
-	/**
-	 * Returns the class name local service.
-	 *
-	 * @return the class name local service
-	 */
-	public com.liferay.portal.service.ClassNameLocalService getClassNameLocalService() {
-		return classNameLocalService;
-	}
-
-	/**
-	 * Sets the class name local service.
-	 *
-	 * @param classNameLocalService the class name local service
-	 */
-	public void setClassNameLocalService(
-		com.liferay.portal.service.ClassNameLocalService classNameLocalService) {
-		this.classNameLocalService = classNameLocalService;
-	}
-
-	/**
-	 * Returns the class name remote service.
-	 *
-	 * @return the class name remote service
-	 */
-	public com.liferay.portal.service.ClassNameService getClassNameService() {
-		return classNameService;
-	}
-
-	/**
-	 * Sets the class name remote service.
-	 *
-	 * @param classNameService the class name remote service
-	 */
-	public void setClassNameService(
-		com.liferay.portal.service.ClassNameService classNameService) {
-		this.classNameService = classNameService;
-	}
-
-	/**
-	 * Returns the class name persistence.
-	 *
-	 * @return the class name persistence
-	 */
-	public ClassNamePersistence getClassNamePersistence() {
-		return classNamePersistence;
-	}
-
-	/**
-	 * Sets the class name persistence.
-	 *
-	 * @param classNamePersistence the class name persistence
-	 */
-	public void setClassNamePersistence(
-		ClassNamePersistence classNamePersistence) {
-		this.classNamePersistence = classNamePersistence;
-	}
-
-	/**
-	 * Returns the resource local service.
-	 *
-	 * @return the resource local service
-	 */
-	public com.liferay.portal.service.ResourceLocalService getResourceLocalService() {
-		return resourceLocalService;
-	}
-
-	/**
-	 * Sets the resource local service.
-	 *
-	 * @param resourceLocalService the resource local service
-	 */
-	public void setResourceLocalService(
-		com.liferay.portal.service.ResourceLocalService resourceLocalService) {
-		this.resourceLocalService = resourceLocalService;
-	}
-
-	/**
-	 * Returns the user local service.
-	 *
-	 * @return the user local service
-	 */
-	public com.liferay.portal.service.UserLocalService getUserLocalService() {
-		return userLocalService;
-	}
-
-	/**
-	 * Sets the user local service.
-	 *
-	 * @param userLocalService the user local service
-	 */
-	public void setUserLocalService(
-		com.liferay.portal.service.UserLocalService userLocalService) {
-		this.userLocalService = userLocalService;
-	}
-
-	/**
-	 * Returns the user remote service.
-	 *
-	 * @return the user remote service
-	 */
-	public com.liferay.portal.service.UserService getUserService() {
-		return userService;
-	}
-
-	/**
-	 * Sets the user remote service.
-	 *
-	 * @param userService the user remote service
-	 */
-	public void setUserService(
-		com.liferay.portal.service.UserService userService) {
-		this.userService = userService;
-	}
-
-	/**
-	 * Returns the user persistence.
-	 *
-	 * @return the user persistence
-	 */
-	public UserPersistence getUserPersistence() {
-		return userPersistence;
-	}
-
-	/**
-	 * Sets the user persistence.
-	 *
-	 * @param userPersistence the user persistence
-	 */
-	public void setUserPersistence(UserPersistence userPersistence) {
-		this.userPersistence = userPersistence;
-	}
-
-	/**
-	 * Returns the d d l record local service.
-	 *
-	 * @return the d d l record local service
-	 */
-	public com.liferay.portlet.dynamicdatalists.service.DDLRecordLocalService getDDLRecordLocalService() {
-		return ddlRecordLocalService;
-	}
-
-	/**
-	 * Sets the d d l record local service.
-	 *
-	 * @param ddlRecordLocalService the d d l record local service
-	 */
-	public void setDDLRecordLocalService(
-		com.liferay.portlet.dynamicdatalists.service.DDLRecordLocalService ddlRecordLocalService) {
-		this.ddlRecordLocalService = ddlRecordLocalService;
-	}
-
-	/**
-	 * Returns the d d l record remote service.
-	 *
-	 * @return the d d l record remote service
-	 */
-	public com.liferay.portlet.dynamicdatalists.service.DDLRecordService getDDLRecordService() {
-		return ddlRecordService;
-	}
-
-	/**
-	 * Sets the d d l record remote service.
-	 *
-	 * @param ddlRecordService the d d l record remote service
-	 */
-	public void setDDLRecordService(
-		com.liferay.portlet.dynamicdatalists.service.DDLRecordService ddlRecordService) {
-		this.ddlRecordService = ddlRecordService;
-	}
-
-	/**
-	 * Returns the d d l record persistence.
-	 *
-	 * @return the d d l record persistence
-	 */
-	public DDLRecordPersistence getDDLRecordPersistence() {
-		return ddlRecordPersistence;
-	}
-
-	/**
-	 * Sets the d d l record persistence.
-	 *
-	 * @param ddlRecordPersistence the d d l record persistence
-	 */
-	public void setDDLRecordPersistence(
-		DDLRecordPersistence ddlRecordPersistence) {
-		this.ddlRecordPersistence = ddlRecordPersistence;
-	}
-
-	public void afterPropertiesSet() {
-		Class<?> clazz = getClass();
-
-		_classLoader = clazz.getClassLoader();
-	}
-
-	public void destroy() {
-	}
-
-	/**
-	 * Returns the Spring bean ID for this bean.
-	 *
-	 * @return the Spring bean ID for this bean
-	 */
-	@Override
-	public String getBeanIdentifier() {
-		return _beanIdentifier;
-	}
-
-	/**
-	 * Sets the Spring bean ID for this bean.
-	 *
-	 * @param beanIdentifier the Spring bean ID for this bean
-	 */
-	@Override
-	public void setBeanIdentifier(String beanIdentifier) {
-		_beanIdentifier = beanIdentifier;
-	}
-
-	@Override
-	public Object invokeMethod(String name, String[] parameterTypes,
-		Object[] arguments) throws Throwable {
-		Thread currentThread = Thread.currentThread();
-
-		ClassLoader contextClassLoader = currentThread.getContextClassLoader();
-
-		if (contextClassLoader != _classLoader) {
-			currentThread.setContextClassLoader(_classLoader);
-		}
-
-		try {
-			return _clpInvoker.invokeMethod(name, parameterTypes, arguments);
-		}
-		finally {
-			if (contextClassLoader != _classLoader) {
-				currentThread.setContextClassLoader(contextClassLoader);
-			}
-		}
-	}
-
-	/**
-	 * Performs an SQL query.
-	 *
-	 * @param sql the sql query
-	 */
-	protected void runSQL(String sql) throws SystemException {
-		try {
-			DataSource dataSource = InfrastructureUtil.getDataSource();
-
-			SqlUpdate sqlUpdate = SqlUpdateFactoryUtil.getSqlUpdate(dataSource,
-					sql, new int[0]);
-
-			sqlUpdate.update();
-		}
-		catch (Exception e) {
-			throw new SystemException(e);
-		}
-	}
-
-	@BeanReference(type = com.liferay.mobilewidgets.service.MobileWidgetsAssetEntryService.class)
-	protected com.liferay.mobilewidgets.service.MobileWidgetsAssetEntryService mobileWidgetsAssetEntryService;
-	@BeanReference(type = com.liferay.mobilewidgets.service.MobileWidgetsDDLRecordService.class)
-	protected com.liferay.mobilewidgets.service.MobileWidgetsDDLRecordService mobileWidgetsDDLRecordService;
-	@BeanReference(type = com.liferay.mobilewidgets.service.MobileWidgetsUserService.class)
-	protected com.liferay.mobilewidgets.service.MobileWidgetsUserService mobileWidgetsUserService;
-	@BeanReference(type = com.liferay.counter.service.CounterLocalService.class)
-	protected com.liferay.counter.service.CounterLocalService counterLocalService;
-	@BeanReference(type = com.liferay.portal.service.ClassNameLocalService.class)
-	protected com.liferay.portal.service.ClassNameLocalService classNameLocalService;
-	@BeanReference(type = com.liferay.portal.service.ClassNameService.class)
-	protected com.liferay.portal.service.ClassNameService classNameService;
-	@BeanReference(type = ClassNamePersistence.class)
-	protected ClassNamePersistence classNamePersistence;
-	@BeanReference(type = com.liferay.portal.service.ResourceLocalService.class)
-	protected com.liferay.portal.service.ResourceLocalService resourceLocalService;
-	@BeanReference(type = com.liferay.portal.service.UserLocalService.class)
-	protected com.liferay.portal.service.UserLocalService userLocalService;
-	@BeanReference(type = com.liferay.portal.service.UserService.class)
-	protected com.liferay.portal.service.UserService userService;
-	@BeanReference(type = UserPersistence.class)
-	protected UserPersistence userPersistence;
-	@BeanReference(type = com.liferay.portlet.dynamicdatalists.service.DDLRecordLocalService.class)
-	protected com.liferay.portlet.dynamicdatalists.service.DDLRecordLocalService ddlRecordLocalService;
-	@BeanReference(type = com.liferay.portlet.dynamicdatalists.service.DDLRecordService.class)
-	protected com.liferay.portlet.dynamicdatalists.service.DDLRecordService ddlRecordService;
-	@BeanReference(type = DDLRecordPersistence.class)
-	protected DDLRecordPersistence ddlRecordPersistence;
-	private String _beanIdentifier;
-	private ClassLoader _classLoader;
-	private MobileWidgetsDDLRecordServiceClpInvoker _clpInvoker = new MobileWidgetsDDLRecordServiceClpInvoker();
-}
