--- webs/mobile-widgets-web/docroot/WEB-INF/src/com/liferay/mobilewidgets/service/base/MobileWidgetsDDLRecordServiceClpInvoker.java
@@ -1,125 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.mobilewidgets.service.base;
-
-import com.liferay.mobilewidgets.service.MobileWidgetsDDLRecordServiceUtil;
-
-import java.util.Arrays;
-
-/**
- * @author José Manuel Navarro
- * @generated
- */
-public class MobileWidgetsDDLRecordServiceClpInvoker {
-	public MobileWidgetsDDLRecordServiceClpInvoker() {
-		_methodName30 = "getBeanIdentifier";
-
-		_methodParameterTypes30 = new String[] {  };
-
-		_methodName31 = "setBeanIdentifier";
-
-		_methodParameterTypes31 = new String[] { "java.lang.String" };
-
-		_methodName34 = "getDDLRecord";
-
-		_methodParameterTypes34 = new String[] { "long", "java.util.Locale" };
-
-		_methodName35 = "getDDLRecords";
-
-		_methodParameterTypes35 = new String[] {
-				"long", "java.util.Locale", "int", "int"
-			};
-
-		_methodName36 = "getDDLRecords";
-
-		_methodParameterTypes36 = new String[] {
-				"long", "long", "java.util.Locale", "int", "int"
-			};
-
-		_methodName37 = "getDDLRecordsCount";
-
-		_methodParameterTypes37 = new String[] { "long" };
-
-		_methodName38 = "getDDLRecordsCount";
-
-		_methodParameterTypes38 = new String[] { "long", "long" };
-	}
-
-	public Object invokeMethod(String name, String[] parameterTypes,
-		Object[] arguments) throws Throwable {
-		if (_methodName30.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes30, parameterTypes)) {
-			return MobileWidgetsDDLRecordServiceUtil.getBeanIdentifier();
-		}
-
-		if (_methodName31.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes31, parameterTypes)) {
-			MobileWidgetsDDLRecordServiceUtil.setBeanIdentifier((java.lang.String)arguments[0]);
-
-			return null;
-		}
-
-		if (_methodName34.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes34, parameterTypes)) {
-			return MobileWidgetsDDLRecordServiceUtil.getDDLRecord(((Long)arguments[0]).longValue(),
-				(java.util.Locale)arguments[1]);
-		}
-
-		if (_methodName35.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes35, parameterTypes)) {
-			return MobileWidgetsDDLRecordServiceUtil.getDDLRecords(((Long)arguments[0]).longValue(),
-				(java.util.Locale)arguments[1],
-				((Integer)arguments[2]).intValue(),
-				((Integer)arguments[3]).intValue());
-		}
-
-		if (_methodName36.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes36, parameterTypes)) {
-			return MobileWidgetsDDLRecordServiceUtil.getDDLRecords(((Long)arguments[0]).longValue(),
-				((Long)arguments[1]).longValue(),
-				(java.util.Locale)arguments[2],
-				((Integer)arguments[3]).intValue(),
-				((Integer)arguments[4]).intValue());
-		}
-
-		if (_methodName37.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes37, parameterTypes)) {
-			return MobileWidgetsDDLRecordServiceUtil.getDDLRecordsCount(((Long)arguments[0]).longValue());
-		}
-
-		if (_methodName38.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes38, parameterTypes)) {
-			return MobileWidgetsDDLRecordServiceUtil.getDDLRecordsCount(((Long)arguments[0]).longValue(),
-				((Long)arguments[1]).longValue());
-		}
-
-		throw new UnsupportedOperationException();
-	}
-
-	private String _methodName30;
-	private String[] _methodParameterTypes30;
-	private String _methodName31;
-	private String[] _methodParameterTypes31;
-	private String _methodName34;
-	private String[] _methodParameterTypes34;
-	private String _methodName35;
-	private String[] _methodParameterTypes35;
-	private String _methodName36;
-	private String[] _methodParameterTypes36;
-	private String _methodName37;
-	private String[] _methodParameterTypes37;
-	private String _methodName38;
-	private String[] _methodParameterTypes38;
-}
