--- webs/mobile-widgets-web/docroot/WEB-INF/src/com/liferay/mobilewidgets/service/base/MobileWidgetsUserServiceClpInvoker.java
@@ -1,92 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.mobilewidgets.service.base;
-
-import com.liferay.mobilewidgets.service.MobileWidgetsUserServiceUtil;
-
-import java.util.Arrays;
-
-/**
- * @author José Manuel Navarro
- * @generated
- */
-public class MobileWidgetsUserServiceClpInvoker {
-	public MobileWidgetsUserServiceClpInvoker() {
-		_methodName30 = "getBeanIdentifier";
-
-		_methodParameterTypes30 = new String[] {  };
-
-		_methodName31 = "setBeanIdentifier";
-
-		_methodParameterTypes31 = new String[] { "java.lang.String" };
-
-		_methodName34 = "sendPasswordByEmailAddress";
-
-		_methodParameterTypes34 = new String[] { "long", "java.lang.String" };
-
-		_methodName35 = "sendPasswordByScreenName";
-
-		_methodParameterTypes35 = new String[] { "long", "java.lang.String" };
-
-		_methodName36 = "sendPasswordByUserId";
-
-		_methodParameterTypes36 = new String[] { "long" };
-	}
-
-	public Object invokeMethod(String name, String[] parameterTypes,
-		Object[] arguments) throws Throwable {
-		if (_methodName30.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes30, parameterTypes)) {
-			return MobileWidgetsUserServiceUtil.getBeanIdentifier();
-		}
-
-		if (_methodName31.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes31, parameterTypes)) {
-			MobileWidgetsUserServiceUtil.setBeanIdentifier((java.lang.String)arguments[0]);
-
-			return null;
-		}
-
-		if (_methodName34.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes34, parameterTypes)) {
-			return MobileWidgetsUserServiceUtil.sendPasswordByEmailAddress(((Long)arguments[0]).longValue(),
-				(java.lang.String)arguments[1]);
-		}
-
-		if (_methodName35.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes35, parameterTypes)) {
-			return MobileWidgetsUserServiceUtil.sendPasswordByScreenName(((Long)arguments[0]).longValue(),
-				(java.lang.String)arguments[1]);
-		}
-
-		if (_methodName36.equals(name) &&
-				Arrays.deepEquals(_methodParameterTypes36, parameterTypes)) {
-			return MobileWidgetsUserServiceUtil.sendPasswordByUserId(((Long)arguments[0]).longValue());
-		}
-
-		throw new UnsupportedOperationException();
-	}
-
-	private String _methodName30;
-	private String[] _methodParameterTypes30;
-	private String _methodName31;
-	private String[] _methodParameterTypes31;
-	private String _methodName34;
-	private String[] _methodParameterTypes34;
-	private String _methodName35;
-	private String[] _methodParameterTypes35;
-	private String _methodName36;
-	private String[] _methodParameterTypes36;
-}
