--- webs/mobile-widgets-web/docroot/WEB-INF/src/com/liferay/mobilewidgets/service/http/MobileWidgetsAssetEntryService_deploy.wsdd
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-
-<deployment xmlns="http://xml.apache.org/axis/wsdd/" xmlns:java="http://xml.apache.org/axis/wsdd/providers/java">
-	<service name="Plugin_MobileWidgets_MobileWidgetsAssetEntryService" provider="java:RPC" style="rpc" use="encoded">
-		<arrayMapping xmlns:ns="urn:http.service.mobilewidgets.liferay.com" xmlns:cmp-ns="http://www.w3.org/2001/XMLSchema" qname="ns:ArrayOfArrayOf_xsd_long" type="java:long[][]" innerType="cmp-ns:long" encodingStyle="http://schemas.xmlsoap.org/soap/encoding/"></arrayMapping>
-		<arrayMapping xmlns:ns="urn:http.service.mobilewidgets.liferay.com" xmlns:cmp-ns="http://www.w3.org/2001/XMLSchema" qname="ns:ArrayOf_xsd_long" type="java:long[]" innerType="cmp-ns:long" encodingStyle="http://schemas.xmlsoap.org/soap/encoding/"></arrayMapping>
-		<typeMapping xmlns:ns="http://persistence.service.asset.portlet.liferay.com" qname="ns:AssetEntryQuery" type="java:com.liferay.portlet.asset.service.persistence.AssetEntryQuery" serializer="org.apache.axis.encoding.ser.BeanSerializerFactory" deserializer="org.apache.axis.encoding.ser.BeanDeserializerFactory" encodingStyle="http://schemas.xmlsoap.org/soap/encoding/"></typeMapping>
-		<typeMapping xmlns:ns="http://xml.apache.org/xml-soap" qname="ns:mapItem" type="java:org.apache.xml.xml_soap.MapItem" serializer="org.apache.axis.encoding.ser.BeanSerializerFactory" deserializer="org.apache.axis.encoding.ser.BeanDeserializerFactory" encodingStyle="http://schemas.xmlsoap.org/soap/encoding/"></typeMapping>
-		<operation xmlns:operNS="urn:http.service.mobilewidgets.liferay.com" xmlns:rtns="http://www.w3.org/2001/XMLSchema" name="getAssetEntries" qname="operNS:getAssetEntries" returnQName="getAssetEntriesReturn" returnType="rtns:string" soapAction="">
-			<parameter xmlns:tns="http://persistence.service.asset.portlet.liferay.com" qname="assetEntryQuery" type="tns:AssetEntryQuery"></parameter>
-			<parameter xmlns:tns="http://www.w3.org/2001/XMLSchema" qname="locale" type="tns:string"></parameter>
-		</operation>
-		<parameter name="allowedMethods" value="getAssetEntries" />
-		<parameter name="className" value="com.liferay.mobilewidgets.service.http.MobileWidgetsAssetEntryServiceSoap" />
-		<parameter name="schemaUnqualified" value="http://persistence.service.asset.portlet.liferay.com,http://xml.apache.org/xml-soap,urn:http.service.mobilewidgets.liferay.com" />
-		<parameter name="typeMappingVersion" value="1.2" />
-		<parameter name="wsdlPortType" value="MobileWidgetsAssetEntryServiceSoap" />
-		<parameter name="wsdlServiceElement" value="MobileWidgetsAssetEntryServiceSoapService" />
-		<parameter name="wsdlServicePort" value="Plugin_MobileWidgets_MobileWidgetsAssetEntryService" />
-		<parameter name="wsdlTargetNamespace" value="urn:http.service.mobilewidgets.liferay.com" />
-	</service>
-</deployment>
