--- webs/mobile-widgets-web/docroot/WEB-INF/src/com/liferay/mobilewidgets/service/http/MobileWidgetsAssetEntryService_undeploy.wsdd
@@ -1,5 +1,0 @@
-<?xml version="1.0"?>
-
-<undeployment xmlns="http://xml.apache.org/axis/wsdd/">
-	<service name="Plugin_MobileWidgets_MobileWidgetsAssetEntryService" />
-</undeployment>
