--- webs/mobile-widgets-web/docroot/WEB-INF/src/com/liferay/mobilewidgets/service/http/MobileWidgetsUserServiceSoap.java
@@ -1,99 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.mobilewidgets.service.http;
-
-import com.liferay.mobilewidgets.service.MobileWidgetsUserServiceUtil;
-
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-
-import java.rmi.RemoteException;
-
-/**
- * Provides the SOAP utility for the
- * {@link com.liferay.mobilewidgets.service.MobileWidgetsUserServiceUtil} service utility. The
- * static methods of this class calls the same methods of the service utility.
- * However, the signatures are different because it is difficult for SOAP to
- * support certain types.
- *
- * <p>
- * The benefits of using the SOAP utility is that it is cross platform
- * compatible. SOAP allows different languages like Java, .NET, C++, PHP, and
- * even Perl, to call the generated services. One drawback of SOAP is that it is
- * slow because it needs to serialize all calls into a text format (XML).
- * </p>
- *
- * <p>
- * You can see a list of services at http://localhost:8080/api/axis. Set the
- * property <b>axis.servlet.hosts.allowed</b> in portal.properties to configure
- * security.
- * </p>
- *
- * <p>
- * The SOAP utility is only generated for remote services.
- * </p>
- *
- * @author José Manuel Navarro
- * @see MobileWidgetsUserServiceHttp
- * @see com.liferay.mobilewidgets.service.MobileWidgetsUserServiceUtil
- * @generated
- */
-public class MobileWidgetsUserServiceSoap {
-	public static boolean sendPasswordByEmailAddress(long companyId,
-		java.lang.String emailAddress) throws RemoteException {
-		try {
-			boolean returnValue = MobileWidgetsUserServiceUtil.sendPasswordByEmailAddress(companyId,
-					emailAddress);
-
-			return returnValue;
-		}
-		catch (Exception e) {
-			_log.error(e, e);
-
-			throw new RemoteException(e.getMessage());
-		}
-	}
-
-	public static boolean sendPasswordByScreenName(long companyId,
-		java.lang.String screenName) throws RemoteException {
-		try {
-			boolean returnValue = MobileWidgetsUserServiceUtil.sendPasswordByScreenName(companyId,
-					screenName);
-
-			return returnValue;
-		}
-		catch (Exception e) {
-			_log.error(e, e);
-
-			throw new RemoteException(e.getMessage());
-		}
-	}
-
-	public static boolean sendPasswordByUserId(long userId)
-		throws RemoteException {
-		try {
-			boolean returnValue = MobileWidgetsUserServiceUtil.sendPasswordByUserId(userId);
-
-			return returnValue;
-		}
-		catch (Exception e) {
-			_log.error(e, e);
-
-			throw new RemoteException(e.getMessage());
-		}
-	}
-
-	private static Log _log = LogFactoryUtil.getLog(MobileWidgetsUserServiceSoap.class);
-}
