--- webs/mobile-widgets-web/docroot/WEB-INF/src/com/liferay/mobilewidgets/service/http/MobileWidgetsUserService_deploy.wsdd
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-
-<deployment xmlns="http://xml.apache.org/axis/wsdd/" xmlns:java="http://xml.apache.org/axis/wsdd/providers/java">
-	<service name="Plugin_MobileWidgets_MobileWidgetsUserService" provider="java:RPC" style="rpc" use="encoded">
-		<operation xmlns:operNS="urn:http.service.mobilewidgets.liferay.com" xmlns:rtns="http://www.w3.org/2001/XMLSchema" name="sendPasswordByEmailAddress" qname="operNS:sendPasswordByEmailAddress" returnQName="sendPasswordByEmailAddressReturn" returnType="rtns:boolean" soapAction="">
-			<parameter xmlns:tns="http://www.w3.org/2001/XMLSchema" qname="companyId" type="tns:long"></parameter>
-			<parameter xmlns:tns="http://www.w3.org/2001/XMLSchema" qname="emailAddress" type="tns:string"></parameter>
-		</operation>
-		<operation xmlns:operNS="urn:http.service.mobilewidgets.liferay.com" xmlns:rtns="http://www.w3.org/2001/XMLSchema" name="sendPasswordByScreenName" qname="operNS:sendPasswordByScreenName" returnQName="sendPasswordByScreenNameReturn" returnType="rtns:boolean" soapAction="">
-			<parameter xmlns:tns="http://www.w3.org/2001/XMLSchema" qname="companyId" type="tns:long"></parameter>
-			<parameter xmlns:tns="http://www.w3.org/2001/XMLSchema" qname="screenName" type="tns:string"></parameter>
-		</operation>
-		<operation xmlns:operNS="urn:http.service.mobilewidgets.liferay.com" xmlns:rtns="http://www.w3.org/2001/XMLSchema" name="sendPasswordByUserId" qname="operNS:sendPasswordByUserId" returnQName="sendPasswordByUserIdReturn" returnType="rtns:boolean" soapAction="">
-			<parameter xmlns:tns="http://www.w3.org/2001/XMLSchema" qname="userId" type="tns:long"></parameter>
-		</operation>
-		<parameter name="allowedMethods" value="sendPasswordByEmailAddress sendPasswordByScreenName sendPasswordByUserId" />
-		<parameter name="className" value="com.liferay.mobilewidgets.service.http.MobileWidgetsUserServiceSoap" />
-		<parameter name="typeMappingVersion" value="1.2" />
-		<parameter name="wsdlPortType" value="MobileWidgetsUserServiceSoap" />
-		<parameter name="wsdlServiceElement" value="MobileWidgetsUserServiceSoapService" />
-		<parameter name="wsdlServicePort" value="Plugin_MobileWidgets_MobileWidgetsUserService" />
-		<parameter name="wsdlTargetNamespace" value="urn:http.service.mobilewidgets.liferay.com" />
-	</service>
-</deployment>
