+++ webs/mobile-widgets-web/docroot/WEB-INF/src/service.properties
@@ -13,8 +13,8 @@
 ##
 
     build.namespace=LiferayScreens
-    build.number=3
-    build.date=1431528736279
+    build.number=4
+    build.date=1431529083989
     build.auto.upgrade=true
 
 ##
